feat: move daily task loading to client
